Cleaned up, changed dependencies, added churn
Refactored lots of code to make the project cleaner. Now looking for a single torrent simulation and later multi torrent, instead of one package for each type of simulation.

Replaced the Apache commons CLI dependency with Java Properties config. This is in the standard library and avoids the use of complicated arguments.

Can now add churn to the simulation. Nodes can abandon their download before it is complete, and can decide to leave a torrent after downloading. Leaving the torrent wipes the node clean but leaves bad data in the network.